Integrate add command with the new parser
