+++ src/main/java/seedu/address/logic/commands/todolistcommands/AddTodoCommand.java
@@ -1,15 +1,66 @@
 package seedu.address.logic.commands.todolistcommands;
 
+import static java.util.Objects.requireNonNull;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_DATE;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_PRIORITY;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
+
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
+import seedu.address.model.task.Task;
 
+/**
+ * Adds a task to the todo list.
+ */
 public class AddTodoCommand extends Command {
+
     public static final String COMMAND_WORD = "addtodo";
+
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Adds a task to the todo list. "
+        + "Parameters: "
+        + PREFIX_NAME + "TASK NAME "
+        + PREFIX_TAG + "TAG [OPTIONAL] "
+        + PREFIX_PRIORITY + "PRIORITY [OPTIONAL] "
+        + PREFIX_DATE + "DATE/DEADLINE [OPTIONAL] "
+        + "Example: " + COMMAND_WORD + " "
+        + PREFIX_NAME + "Do Lab Sheet 1 "
+        + PREFIX_TAG + "LAB "
+        + PREFIX_PRIORITY + "HIGH "
+        + PREFIX_DATE + "2020-10-16 ";
+
+    public static final String MESSAGE_SUCCESS = "New task added: %1$s";
+    public static final String MESSAGE_DUPLICATE_TASK = "This task already exists in the todo list";
+
+    private final Task toAdd;
+
+    /**
+     * Creates an AddTaskCommand to add the specified {@code Task}
+     */
+    public AddTodoCommand(Task task) {
+        requireNonNull(task);
+        toAdd = task;
+    }
+
     @Override
     public CommandResult execute(Model model) throws CommandException {
-        return null;
+        requireNonNull(model);
+
+        if (model.hasTask(toAdd)) {
+            throw new CommandException(MESSAGE_DUPLICATE_TASK);
+        }
+
+        model.addTask(toAdd);
+        return new CommandResult(String.format(MESSAGE_SUCCESS, toAdd));
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+            || (other instanceof AddTodoCommand // instanceof handles nulls
+            && toAdd.equals(((AddTodoCommand) other).toAdd));
     }
 
     @Override
@@ -17,3 +68,4 @@ public class AddTodoCommand extends Command {
         return false;
     }
 }
+
